requirements.txt
beautifulsoup4==4.12.2
deep_translator==1.10.1
Django==4.1.7
gTTS==2.3.1
mock==5.0.2
Pillow==9.5.0
Pillow==9.5.0
pytest==7.2.1
python-dotenv==1.0.0
Requests==2.28.2
schema==0.7.5
selenium==4.9.0
webdriver_manager==3.8.6
